small fixes on load and save db

+++ database.py
@@ -13,12 +13,12 @@ db_file = 'database.sqlite'
 
 remote_db_dir = "/home/site/wwwroot/db_dir/"
 
-if env in os.environ:
-    if not os.path.exists(remote_db_dir):
-        os.makedirs(remote_db_dir)
-        print(f"Directory {remote_db_dir} has been created.")
-    else:
-        print(f"Directory {remote_db_dir} already exists.")
+# if env in os.environ:
+#     if not os.path.exists(remote_db_dir):
+#         os.makedirs(remote_db_dir)
+#         print(f"Directory {remote_db_dir} has been created.")
+#     else:
+#         print(f"Directory {remote_db_dir} already exists.")
 
 
 remote_db = f"{remote_db_dir}{db_file}"
@@ -35,19 +35,29 @@ local_db = db_file
 # copy data
 def load_db():
     if env in os.environ:
-        try:
+        if not os.path.exists(remote_db_dir):
+            os.makedirs(remote_db_dir)
+            print(f"Directory {remote_db_dir} has been created.")
+        else:
+            print(f"Directory {remote_db_dir} already exists.")
+        if os.path.exists(remote_db_dir):
             shutil.copy(remote_db, local_db)
             print(f"Copyed {remote_db} to {local_db}")
-        except:
-            print("Source file not found")
+        else:
+            print("Remote directory not found")
 
 def save_db():
     if env in os.environ:
-        try:
+        if not os.path.exists(remote_db_dir):
+            os.makedirs(remote_db_dir)
+            print(f"Directory {remote_db_dir} has been created.")
+        else:
+            print(f"Directory {remote_db_dir} already exists.")
+        if os.path.exists(remote_db_dir):
             shutil.copy(local_db, remote_db)
             print(f"Copyed {local_db} to {remote_db}")
-        except:
-            print("Source file not found")
+        else:
+            print("Remote directory not found")
 
 
 db_connection = f'sqlite:///{local_db}'
